Refactor indra db rest client into proper processor.
